Add some string builtins
Started off with some low-hanging fruits.

Others will be more complicated as we'd need to extend the argument
type checks to work with collections of certain types, or they don't
have an equivalent built-in in clojure. But hey, it's a start :)

Note that I skipped doing args type check tests here as we're planning
to move that into the evaluator.

Signed-off-by: Anders Eknert <anders@eknert.com>