Allow builtins resolver to be provided
This solves the opa.runtime test which failed, since
the OPA tests have mocked this to return an empty map..
but more importantly, it will allow us and others to
easily provide their set of builtins, or even swap out
the ones that we provide.

Touching some code that isn't my domain here, so if this
feels hacky, let me know.

Signed-off-by: Anders Eknert <anders@eknert.com>