Move type checking to builtin function evaluation
And use the type definitions provided in the plan for each
built-in function. This allows us to decouple basic validation
of arguments from the function definitions, removing the need
to use metadata in order to provide the types of the arguments.

Functions are of course still free to perform extend type checks,
like e.g. ensure that a provided "number" is an integer and not a
float.

Signed-off-by: Anders Eknert <anders@eknert.com>